Added cookie utilities.
request-cookies - returns a list of parsed cookies
request-cookie-ref - extract info of named cookie
response-cookie-add! - adds a cookie to the response
response-cookie-delete! - deletes a cookie from the response

NB: response-cookie-delete! merely omit set-cookie header
for the named cookie; it does not delete client-side cookie.
To delete the client-side cookie, set the null value to it.