Changed terminate-server-loop to use channels instead of conditions
In actual use case, calling terminate-server-loop from request
handler is uncommon.  Instead of forcing the function to be called
within the handler, we employed an explicitly created communication
channel to request the server loop to quit.  This way, it can be
used in the case we have multiple server loops running in separate
threads, and the controlling thread can terminate each of them as
needed.