For forwarded situation, we need default setting of port number
This commit is provisional.  It seems that there are a few conventions.
In some, the proxy's host and port is passed to x-forwarded-host
and x-forwarded-port, but it is not clear what x-forwarded-port has
when the client used the default port (thus no explit port in the uri).
Some use host:port format in x-forwarded-host and do not use
x-forwarded-port.
For now, we fallback to 80 when we don't see x-forwarded-port, but
this breaks when proxy is running on other than 80.