Start introducing path-component match
We allow PATTERN to be a list of path components, where a string element
matches corresponding path component, and a symbol element
matches anything and saves the matched component with the symbol as its name.

The transition is not complete--the new spec conflicts with optional METHOD
list, which needs to be taken care of.