Emit DOCTYPE when returning sxml/html content
https://github.com/shirok/Gauche-makiki/issues/9

sxml:sxml->html does not add DOCTYPE by default.  Now that HTML5 is
universal, it's better to provide it by default.

We could make it customizable with a special value in substitution
alist, e.g. (respond/ok '(sxml <sxml> (DOCTYPE ...))'.  Not sure
how to provide parameters, though. so it's for future implemtation.